Split config and storage files to clarify structure
