+++ storage/config.go
@@ -1,0 +1,7 @@
+package storage
+
+// Config is an interface of factory method for Storage
+type Config interface {
+	// NewStorage returns a new instance of Storage.
+	NewStorage() (Storage, error)
+}
