+++ storage/local/storage.go
@@ -8,21 +8,6 @@ import (
 	"github.com/minamijoyo/tfmigrate/storage"
 )
 
-// Config is a config for local storage.
-type Config struct {
-	// Path to a migration history file. Relative to the current working directory.
-	Path string `hcl:"path"`
-}
-
-// Config implements a storage.Config.
-var _ storage.Config = (*Config)(nil)
-
-// NewStorage returns a new instance of storage.Storage.
-func (c *Config) NewStorage() (storage.Storage, error) {
-	s := NewStorage(c.Path)
-	return s, nil
-}
-
 // Storage is a storage.Storage implementation for local file.
 // This was originally intended for debugging purposes, but it can also be used
 // as a workaround if Storage doesn't support your cloud provider.
