+++ storage/storage.go
@@ -15,9 +15,3 @@ type Storage interface {
 	// an empty array instead of an error.
 	Read(ctx context.Context) ([]byte, error)
 }
-
-// Config is an interface of factory method for Storage
-type Config interface {
-	// NewStorage returns a new instance of Storage.
-	NewStorage() (Storage, error)
-}
